Convert all .unwraps()s that don't change signatures nor tests
Unwraps that don't change signatures are now converted to being
safer. However this is not everything and further work is
required. Namely tests are not converted, this would normally be fine,
but tests are exported to Elixir and thus have a chance to crash the
VM.

Further functions like:

tests::create_an_action

Have non Result outputs, meaning greater refactoring is required to
handle these.